refactor: Analyze 서비스 리팩토링
